Added File type. Removed NodeInfo type
